+++ internal/observability/middleware.go
@@ -22,7 +22,6 @@ package observability
 
 import (
 	"context"
-	"fmt"
 	"sync"
 
 	"go.uber.org/net/metrics"
@@ -91,7 +90,11 @@ type Config struct {
 	// Scope to which metrics are emitted.
 	Scope *metrics.Scope
 
-	// Extracts request-scoped information from the context for logging.
+	// MetricTagsBlocklist of metric tags being suppressed from being tagged on
+	// metrics emitted by the middleware.
+	MetricTagsBlocklist []string
+
+	// ContextExtractor Extracts request-scoped information from the context for logging.
 	ContextExtractor ContextExtractor
 
 	// Levels specify log levels for various classes of requests.
@@ -118,19 +121,33 @@ type DirectionalLevelsConfig struct {
 	// This includes low-level network errors, TChannel error frames, etc.
 	//
 	// Defaults to ErrorLevel.
+	// Deprecated in favor of ServerError and ClientError.
 	Failure *zapcore.Level
 
 	// Log level used to log calls that failed with an application error.
 	// All Thrift exceptions are considered application errors.
 	//
 	// Defaults to ErrorLevel.
+	// Deprecated in favor of ServerError and ClientError.
 	ApplicationError *zapcore.Level
+
+	// Log level used to log calls that failed with an server error.
+	//
+	// Defaults to ErrorLevel.
+	ServerError *zapcore.Level
+
+	// Log level used to log calls that failed with an client error.
+	// All Thrift exceptions are considered application errors if
+	// there are not annotated with the option rpc.code.
+	//
+	// Defaults to ErrorLevel.
+	ClientError *zapcore.Level
 }
 
 // NewMiddleware constructs an observability middleware with the provided
 // configuration.
 func NewMiddleware(cfg Config) *Middleware {
-	m := &Middleware{newGraph(cfg.Scope, cfg.Logger, cfg.ContextExtractor)}
+	m := &Middleware{newGraph(cfg.Scope, cfg.Logger, cfg.ContextExtractor, cfg.MetricTagsBlocklist)}
 
 	// Apply the default levels
 	applyLogLevelsConfig(&m.graph.inboundLevels, &cfg.Levels.Default)
@@ -147,10 +164,18 @@ func applyLogLevelsConfig(dst *levels, src *DirectionalLevelsConfig) {
 		dst.success = *src.Success
 	}
 	if level := src.Failure; level != nil {
+		dst.useApplicationErrorFailureLevels = true
 		dst.failure = *src.Failure
 	}
 	if level := src.ApplicationError; level != nil {
+		dst.useApplicationErrorFailureLevels = true
 		dst.applicationError = *src.ApplicationError
+	}
+	if level := src.ServerError; level != nil {
+		dst.serverError = *src.ServerError
+	}
+	if level := src.ClientError; level != nil {
+		dst.clientError = *src.ClientError
 	}
 }
 
@@ -183,6 +208,8 @@ func (m *Middleware) Handle(ctx context.Context, req *transport.Request, w trans
 // Call implements middleware.UnaryOutbound.
 func (m *Middleware) Call(ctx context.Context, req *transport.Request, out transport.UnaryOutbound) (*transport.Response, error) {
 	call := m.graph.begin(ctx, transport.Unary, _directionOutbound, req)
+	defer m.handlePanicForCall(call, transport.Unary)
+
 	res, err := out.Call(ctx, req)
 
 	isApplicationError := false
@@ -207,6 +234,8 @@ func (m *Middleware) Call(ctx context.Context, req *transport.Request, out trans
 // HandleOneway implements middleware.OnewayInbound.
 func (m *Middleware) HandleOneway(ctx context.Context, req *transport.Request, h transport.OnewayHandler) error {
 	call := m.graph.begin(ctx, transport.Oneway, _directionInbound, req)
+	defer m.handlePanicForCall(call, transport.Oneway)
+
 	err := h.HandleOneway(ctx, req)
 	call.End(callResult{err: err, requestSize: req.BodySize})
 	return err
@@ -215,6 +244,8 @@ func (m *Middleware) HandleOneway(ctx context.Context, req *transport.Request, h
 // CallOneway implements middleware.OnewayOutbound.
 func (m *Middleware) CallOneway(ctx context.Context, req *transport.Request, out transport.OnewayOutbound) (transport.Ack, error) {
 	call := m.graph.begin(ctx, transport.Oneway, _directionOutbound, req)
+	defer m.handlePanicForCall(call, transport.Oneway)
+
 	ack, err := out.CallOneway(ctx, req)
 	call.End(callResult{err: err, requestSize: req.BodySize})
 	return ack, err
@@ -234,6 +265,13 @@ func (m *Middleware) HandleStream(serverStream *transport.ServerStream, h transp
 // CallStream implements middleware.StreamOutbound.
 func (m *Middleware) CallStream(ctx context.Context, request *transport.StreamRequest, out transport.StreamOutbound) (*transport.ClientStream, error) {
 	call := m.graph.begin(ctx, transport.Streaming, _directionOutbound, request.Meta.ToRequest())
+	defer m.handlePanicForCall(call, transport.Streaming)
+
+	// TODO: metrics for outbound stream and inbound stream are not reported
+	// in the same way. HandleStream increases the metric calls by 1 before passing
+	// the request while CallStream does it after passing the request.
+	// In case of panics in CallStream, the metrics calls will not be increased
+	// while it will in HandleStream.
 	clientStream, err := out.CallStream(ctx, request)
 	call.EndStreamHandshakeWithError(err)
 	if err != nil {
@@ -260,15 +298,14 @@ func ctxErrOverride(ctx context.Context, req *transport.Request) (ctxErr error) 
 
 // handlePanicForCall checks for a panic without actually recovering from it
 // it must be called in defer otherwise recover will act as a no-op
-// The only action this method takes is to emit panic metrics
+// The only action this method takes is to emit panic metrics.
 func (m *Middleware) handlePanicForCall(call call, transportType transport.Type) {
 	// We only want to emit panic metrics without actually recovering from it
 	// Actual recovery from a panic happens at top of the stack in transport's Handler Invoker
 	// As this middleware is the one and only one with Metrics responsibility, we just panic again after
-	// checking for panic without actually recovering from it
+	// checking for panic without actually recovering from it.
 	if e := recover(); e != nil {
-		err := fmt.Errorf("panic: %v", e)
-
+		err := yarpcerrors.InternalErrorf("panic %v", e)
 		// Emit only the panic metrics
 		if transportType == transport.Streaming {
 			call.EndStreamWithPanic(err)