+++ CHANGELOG.md
@@ -4,9 +4,45 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.53.2] - 2021-04-16
+### Removed
+- Disable `rpc-caller-procedure` header temporarily by stopping the `CallerProcedure` propagation.
+### Added
+- gRPC: log header values upon validation error
+
+## [1.53.1] - 2021-03-30
+- v1.53.1 is v1.52.0. v1.53.0 has a backward compatible issue with the new header
+  `rpc-caller-procedure` added in v1.53.0.
+
+## [1.53.0] - 2021-03-12
+### Added
+- gRPC: accept keepalive parameters for gRPC outbound configuration.
+- observability: deprecation of logging levels `applicationError` and `serverError` in favor
+  of `serverError` and `clientError`.
+- The `transport.Request` provides a new field `CallerProcedure`. A header `rpc-caller-procedure`
+  is also propagated for outbound calls. This field represents the origin procedure where the
+  requests was sent.
+### Fixed
+- peer: concurrency issue when using `rand` for the peer strategies `randpeer` and `tworandomchoices`
+
+## [1.52.0] - 2021-02-12
+### Added
+- gRPC: accept keepalive parameters for the outbound gRPC connection.
+- observability: panic metrics are now also reported for these type of requests:
+  - inbound oneway
+  - outbound stream
+  - outbound oneway
+  - outbound unary
+
+## [1.51.0] - 2021-02-04
+### Added
+- observability: Add metric tags blocklist configuration. Allows to stub high cardinality tags emitted
+  in the observability middleware
+
 ## [1.50.0] - 2021-01-22
 ### Added
 - gRPC: accept dialer options in gRPC transportSpec config
+
 ### Fixed
 - Http: fix memory and connection leak in http outbound call handler
 - Observability: error code CodeUnimplemented will be marked as an `application_error`
@@ -1313,7 +1349,12 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
-[1.50.0]: https://github.com/yarpc/yarpc-go/compare/v1.50.0...v1.50.1
+[1.53.2]: https://github.com/yarpc/yarpc-go/compare/v1.53.1...v1.53.2
+[1.53.1]: https://github.com/yarpc/yarpc-go/compare/v1.51.0...v1.52.0
+[1.53.0]: https://github.com/yarpc/yarpc-go/compare/v1.52.0...v1.53.0
+[1.52.0]: https://github.com/yarpc/yarpc-go/compare/v1.51.0...v1.52.0
+[1.51.0]: https://github.com/yarpc/yarpc-go/compare/v1.50.0...v1.51.0
+[1.50.0]: https://github.com/yarpc/yarpc-go/compare/v1.49.1...v1.50.0
 [1.49.1]: https://github.com/yarpc/yarpc-go/compare/v1.49.0...v1.49.1
 [1.49.0]: https://github.com/yarpc/yarpc-go/compare/v1.48.0...v1.49.0
 [1.48.0]: https://github.com/yarpc/yarpc-go/compare/v1.47.2...v1.48.0